+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-27T13:21:49.218Z",
+    "2026-08-31T07:26:08.091Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {