+++ packages/orq-rc/RELEASES.md
@@ -11969,3 +11969,13 @@ Based on:
 - [typescript v4.15.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.6 - packages/orq-rc
+
+## 2026-08-31 07:28:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.2 (2.933.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.7 - packages/orq-rc